gateway-requirements.txt
numpy==1.26.0
Pillow==10.0.1
requests==2.31.0
Flask==3.0.0
gunicorn==21.2.0